requirements.txt
torch>=2.1
jax>=0.4
numpy>=1.26
matplotlib>=3.8
PyYAML>=6.0
tqdm>=4.66
